third_party/blink/web_tests/virtual/relax-dom-valid-names-disabled/fast/dom/Document/createAttributeNS-namespace-err-expected.txt

createAttirbuteNS tests adapted from createElementNS tests attached to webkit bug 16833

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS createAttributeNS(undefined, undefined)
PASS createAttributeNS(null, undefined)
PASS createAttributeNS(undefined, null)
PASS createAttributeNS(null, null)
PASS createAttributeNS(null, ""); threw InvalidCharacterError: Failed to execute 'createAttributeNS' on 'Document': The qualified name provided is empty.
PASS createAttributeNS("", null)
PASS createAttributeNS("", ""); threw InvalidCharacterError: Failed to execute 'createAttributeNS' on 'Document': The qualified name provided is empty.
PASS createAttributeNS(null, "<div>"); threw InvalidCharacterError: Failed to execute 'createAttributeNS' on 'Document': The qualified name provided ('<div>') contains the invalid name-start character '<'.
FAIL createAttributeNS(null, "0div"); expected expected no exception, threw INVALID_CHARACTER_ERR
PASS createAttributeNS(null, "di v"); threw InvalidCharacterError: Failed to execute 'createAttributeNS' on 'Document': The qualified name provided ('di v') contains the invalid character ' '.
FAIL createAttributeNS(null, "di<v"); expected expected no exception, threw INVALID_CHARACTER_ERR
FAIL createAttributeNS(null, "-div"); expected expected no exception, threw INVALID_CHARACTER_ERR
FAIL createAttributeNS(null, ".div"); expected expected no exception, threw INVALID_CHARACTER_ERR
PASS createAttributeNS("http://example.com/", "<div>"); threw InvalidCharacterError: Failed to execute 'createAttributeNS' on 'Document': The qualified name provided ('<div>') contains the invalid name-start character '<'.
FAIL createAttributeNS("http://example.com/", "0div"); expected expected no exception, threw INVALID_CHARACTER_ERR
FAIL createAttributeNS("http://example.com/", "di<v"); expected expected no exception, threw INVALID_CHARACTER_ERR
FAIL createAttributeNS("http://example.com/", "-div"); expected expected no exception, threw INVALID_CHARACTER_ERR
FAIL createAttributeNS("http://example.com/", ".div"); expected expected no exception, threw INVALID_CHARACTER_ERR
PASS createAttributeNS(null, ":div"); threw InvalidCharacterError: Failed to execute 'createAttributeNS' on 'Document': The qualified name provided (':div') has an empty namespace prefix.
PASS createAttributeNS(null, "div:"); threw InvalidCharacterError: Failed to execute 'createAttributeNS' on 'Document': The qualified name provided ('div:') has an empty local name.
PASS createAttributeNS("http://example.com/", ":div"); threw InvalidCharacterError: Failed to execute 'createAttributeNS' on 'Document': The qualified name provided (':div') has an empty namespace prefix.
PASS createAttributeNS("http://example.com/", "div:"); threw InvalidCharacterError: Failed to execute 'createAttributeNS' on 'Document': The qualified name provided ('div:') has an empty local name.
PASS createAttributeNS(null, "d:iv"); threw NamespaceError: Failed to execute 'createAttributeNS' on 'Document': The namespace URI provided ('') is not valid for the qualified name provided ('d:iv').
FAIL createAttributeNS(null, "a:b:c"); expected NAMESPACE_ERR, threw INVALID_CHARACTER_ERR
FAIL createAttributeNS("http://example.com/", "a:b:c"); expected expected no exception, threw INVALID_CHARACTER_ERR
PASS createAttributeNS(null, "a::c"); threw InvalidCharacterError: Failed to execute 'createAttributeNS' on 'Document': The qualified name provided ('a::c') contains multiple colons.
PASS createAttributeNS("http://example.com/", "a::c"); threw InvalidCharacterError: Failed to execute 'createAttributeNS' on 'Document': The qualified name provided ('a::c') contains multiple colons.
FAIL createAttributeNS("http://example.com/", "a:0"); expected expected no exception, threw INVALID_CHARACTER_ERR
FAIL createAttributeNS("http://example.com/", "0:a"); expected expected no exception, threw INVALID_CHARACTER_ERR
PASS createAttributeNS("http://example.com/", "a:_")
FAIL createAttributeNS("http://example.com/", "a:ெ"); expected expected no exception, threw INVALID_CHARACTER_ERR
FAIL createAttributeNS("http://example.com/", "ெ:a"); expected expected no exception, threw INVALID_CHARACTER_ERR
PASS createAttributeNS("http://example.com/", "a:aெ")
PASS createAttributeNS("http://example.com/", "aெ:a")
PASS createAttributeNS("http://example.com/", "xml:test"); binding xml prefix wrong; threw NamespaceError: Failed to execute 'createAttributeNS' on 'Document': The namespace URI provided ('http://example.com/') is not valid for the qualified name provided ('xml:test').
PASS createAttributeNS("http://example.com/", "xmlns:test"); binding xmlns prefix wrong; threw NamespaceError: Failed to execute 'createAttributeNS' on 'Document': The namespace URI provided ('http://example.com/') is not valid for the qualified name provided ('xmlns:test').
PASS createAttributeNS("http://www.w3.org/2000/xmlns/", "x:test"); binding namespace namespace to wrong prefix; threw NamespaceError: Failed to execute 'createAttributeNS' on 'Document': The namespace URI provided ('http://www.w3.org/2000/xmlns/') is not valid for the qualified name provided ('x:test').
PASS createAttributeNS("http://www.w3.org/2000/xmlns/", "xmlns:test")
PASS createAttributeNS("http://www.w3.org/XML/1998/namespace", "xml:test")
PASS createAttributeNS("http://www.w3.org/XML/1998/namespace", "x:test")
PASS createAttributeNS("http://www.w3.org/2000/xmlns/", "xmlns")
PASS createAttributeNS("http://example.com/", "xmlns"); threw NamespaceError: Failed to execute 'createAttributeNS' on 'Document': The namespace URI provided ('http://example.com/') is not valid for the qualified name provided ('xmlns').
PASS successfullyParsed is true

TEST COMPLETE
